requirements/base.txt
asn1crypto==0.24.0
bcrypt==3.1.4
boto==2.48.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.1
cryptography==2.1.4
dj-database-url==0.4.2
dj-email-url==0.0.10
Django==1.11.1
django-coverage-plugin==1.4.2
django-filter==1.0.2
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-storages==1.1.8
djangorestframework==3.6.3
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
paramiko==2.4.0
Pillow==4.3.0
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
storage==0.0.4.3
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.0
